Update README.md
Copper.
•	In third case we will give different categories of rooms which are luxury, ordinary and royal and ask the customer to choose any one that they like.
•	There is also a buffet system for customers who only want to reserve for buffet. 
•	In forth case if customer wants to delete their records they can ask to delete their data/information.
•	In fifth case if the customer only wants to check out the hotel and leave so they can simply press the exit case and leave. 
•	Default case is used in case the user presses a different case that is not mentioned in the main menu.

1.2	Scope of the Project

The program is based on hotel management system. This program is designed to fulfill the requirements that a customer needs in his selection of room. Moreover, the program will facilitate three types of customers: 
 New Customers, Existing Customers and Customers who have been availing our services along the years.  
Our product is the room that is being allotted to our customers. Moreover, the main scope of this program is to make sure that customer gets his desired room, they get their packages and if they want to delete their record, they delete it too.